# guestbook/resources/guestbook-openapi.yml
openapi: 3.0.3
info:
  title: Guestbook - OpenAPI 3.0
  description: >-
    This is a sample Guestbook Server based on the OpenAPI 3.0 specification. 


    Some useful links:

    - [The Guestbook
    repository](https://github.com/morpheus-oss/spring-demos/guestbook)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://guestbook.swagger.io/api/v1
...
paths:
...
  /city:
    get:
      tags:
        - city
      summary: List all cities
      description: List all cities
      operationId: listCities
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
            application/xml:
              schema:
                $ref: '#/components/schemas/City'
      security:
        - guestbook_auth:
            - write:cities
            - read:cities
    post:
      tags:
        - city
      summary: Add a new city
      description: Add a new city
      operationId: addCity
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/City'
          application/xml:
            schema:
              $ref: '#/components/schemas/City'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/City'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
            application/xml:
              schema:
                $ref: '#/components/schemas/City'
      security:
        - guestbook_auth:
            - write:cities
            - read:cities
  /city/{cityId}:
    put:
      tags:
        - city
      summary: Update city, state (name change)
      description: Update city, state (name change)
      operationId: updateCity
      parameters:
        - name: cityId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/City'
          application/xml:
            schema:
              $ref: '#/components/schemas/City'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/City'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
            application/xml:
              schema:
                $ref: '#/components/schemas/City'
      security:
        - guestbook_auth:
            - write:cities
            - read:cities
  /city/{stateId}:
    get:
      tags:
        - city
      summary: Get all cities in the state
      description: Get all cities in the state
      operationId: getCitiesByState
      parameters:
        - name: stateId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
            application/xml:
              schema:
                $ref: '#/components/schemas/City'
        '404':
          description: State not found
      security:
        - guestbook_auth:
            - write:cities
            - read:cities
...
  /state:
    get:
      tags:
        - state
      summary: List all states and union-territories of India
      description: List all states and union-territories of India
      operationId: listStates
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/State'
            application/xml:
              schema:
                $ref: '#/components/schemas/State'
      security:
        - guestbook_auth:
            - write:cities
            - read:cities
  /state/{stateid}:
    get:
      tags: 
        - state
      summary: Get a particular state information
      description: Get a particular state information
      operationId: getState
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/State'
            application/xml:
              schema:
                $ref: '#/components/schemas/State'
      security:
        - guestbook_auth:
            - write:cities
            - read:cities
components:
  schemas:
    State:
      type: object
      properties:
        id:
          type: integer
          example: 5
        name:
          type: string
          example: Assam
    City:
      type: object
      properties:
        name:
          type: string
          example: Bangalore
        state:
          type: string
          example: Karnataka
...
  securitySchemes:
    guestbook_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://guestbook.swagger.io/oauth/authorize
          scopes:
            write:ratings: Add/Modify ratings
            read:ratings: Read your ratings
            write:hotels: Add/Modify hotel details
            read:hotels: Read hotel details
            write:cities: Add/Modify cities/localities
            read:cities: Read all cities/localities
    api_key:
      type: apiKey
      name: api_key
      in: header